- Got proper fonts working in the browser with SFML - Fixed occasional centered font blurriness with SFML (looks like I have to be careful not to draw text at non-integral offsets) - Changed browser item layout a bit so the text isn't overlapping the item image, which could get hard to read at times
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@1177 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/Drawing.cpp
@@ -297,7 +297,7 @@ void Drawing::drawText(string text, int x, int y, rgba_t colour, int font, int a
 		float width = sf_str.GetRect().GetWidth();
 
 		if (alignment == ALIGN_CENTER)
-			sf_str.Move(-(width*0.5), 0.0f);
+			sf_str.Move(-int(width*0.5), 0.0f);
 		else
 			sf_str.Move(-width, 0.0f);
 	}
